+++ PixelProject/Scripts/World/Pixels/Types/SandPixel.h
@@ -30,7 +30,7 @@ class SandPixel final : public BasePixel
 		update_function = static_cast<UpdateFunction>(&SandPixel::PixelUpdate);
 	}
 
-	void PixelUpdate(PixelUpdateResult& data, Uint32& pixel_value)
+	void PixelUpdate(PixelUpdateResult& data, Uint64& pixel_value)
 	{
 		switch (data.Dir()) {
 		case WorldDir::SouthEast:
